+++ __tests__/render.test.js
@@ -1,7 +1,6 @@
 const path = require('path');
 const hbs = require('../dist/build.common.js');
 // const hbs = require('express-hbs-compile');
-const removeSpace = require('./support/removeSpace.js');
 
 test('Any partialsDir, layoutsDir, defaultLayout should be readable', async () => {
   const render = hbs({
@@ -10,8 +9,8 @@ test('Any partialsDir, layoutsDir, defaultLayout should be readable', async () =
     layoutsDir: path.join(__dirname, 'views/layout'),
     defaultLayout: path.join(__dirname, 'views/layout/default.hbs'),
   });
-  const html = await render('index.hbs', {name: 'foo'});
-  expect(removeSpace(html)).toBe('<html><head><title>index</title></head><body><partial>1</partial><subpartial>1</subpartial><subpartial>2</subpartial>Hello,foo</body></html>');
+  const html = await render('basic.hbs', {name: 'foo'});
+  expect(html).toBe('<p>This is a partial view</p><p>Hello, foo</p>');
 });
 
 test('If the partialsDir option is omitted, the default partial template should be loaded', async () => {
@@ -20,8 +19,8 @@ test('If the partialsDir option is omitted, the default partial template should 
     layoutsDir: path.join(__dirname, 'views/layout'),
     defaultLayout: path.join(__dirname, 'views/layout/default.hbs'),
   });
-  const html = await render('index.hbs', {name: 'foo'});
-  expect(removeSpace(html)).toBe('<html><head><title>index</title></head><body><partial>1</partial><subpartial>1</subpartial><subpartial>2</subpartial>Hello,foo</body></html>');
+  const html = await render('basic.hbs', {name: 'foo'});
+  expect(html).toBe('<p>This is a partial view</p><p>Hello, foo</p>');
 });
 
 test('If the layoutsDir option is omitted, the default layout template should be loaded', async () => {
@@ -30,8 +29,8 @@ test('If the layoutsDir option is omitted, the default layout template should be
     partialsDir: path.join(__dirname, 'views/partials'),
     defaultLayout: path.join(__dirname, 'views/layout/default.hbs'),
   });
-  const html = await render('index.hbs', {name: 'foo'});
-  expect(removeSpace(html)).toBe('<html><head><title>index</title></head><body><partial>1</partial><subpartial>1</subpartial><subpartial>2</subpartial>Hello,foo</body></html>');
+  const html = await render('basic.hbs', {name: 'foo'});
+  expect(html).toBe('<p>This is a partial view</p><p>Hello, foo</p>');
 });
 
 test('If the defaultLayout option is omitted, the default layout template should be loaded', async () => {
@@ -40,8 +39,8 @@ test('If the defaultLayout option is omitted, the default layout template should
     partialsDir: path.join(__dirname, 'views/partials'),
     layoutsDir: path.join(__dirname, 'views/layout'),
   });
-  const html = await render('index.hbs', {name: 'foo'});
-  expect(removeSpace(html)).toBe('<html><head><title>index</title></head><body><partial>1</partial><subpartial>1</subpartial><subpartial>2</subpartial>Hello,foo</body></html>');
+  const html = await render('basic.hbs', {name: 'foo'});
+  expect(html).toBe('<p>This is a partial view</p><p>Hello, foo</p>');
 });
 
 test('eq custom helpers should be available', async () => {
@@ -49,5 +48,16 @@ test('eq custom helpers should be available', async () => {
     viewsDir: path.join(__dirname, 'views'),
   });
   const html = await render('eq-helper.hbs', {gender: 1});
-  expect(removeSpace(html)).toBe('<html><head><title>index</title></head><body>Male</body></html>');
+  expect(html).toBe('<p>Male</p>');
+});
+
+test('Custom helpers must work', async () => {
+  const render = hbs({
+    viewsDir: path.join(__dirname, 'views'),
+    helpers: {
+      sayhello: name => `Hello, ${name}`,
+    }
+  });
+  const html = await render('custom-helper.hbs');
+  expect(html).toBe('<p>Hello, Emma</p>');
 });