+++ __tests__/support/removeSpace.js
@@ -1,6 +1,0 @@
-/**
- * Remove space.
- */
-module.exports = str => {
-  return str.replace(/\s+/g, '');
-}