+++ src/interfaces/CompilerOptions.ts
@@ -1,3 +1,5 @@
+import Handlebars from 'handlebars';
+
 /**
  * Compilation options.
  */
@@ -42,4 +44,9 @@ export default interface CompilerOptions {
    * Override the 'block' helper name used in the template.
    */
   blockHelperName?: string;
+
+  /**
+   * A helper accessible from the template. The key is the helper name and the value is the object that will be the helper function.
+   */
+  helpers?: {[key: string]: Handlebars.HelperDelegate};
 }